refactor: streamline analytics data structure and component integration
- Removed the IVotesDistribution interface and associated votesDistribution structure from IPayout and analytics components to simplify data handling.
- Updated PostDetails and PostAnalytics components to fetch and display analytics data more efficiently.
- Enhanced loading state management in PostAnalytics for improved user experience.
- Cleaned up unused imports and adjusted component props for better clarity and performance.
